Moved pattern formatter from spdlog/details to spdlog/
