+++ include/spdlog/logger-inl.h
@@ -9,7 +9,7 @@
 
 #include <spdlog/sinks/sink.h>
 #include <spdlog/details/backtracer.h>
-#include <spdlog/details/pattern_formatter.h>
+#include <spdlog/pattern_formatter.h>
 
 #include <cstdio>
