+++ include/spdlog/sinks/dist_sink.h
@@ -6,7 +6,7 @@
 #include "base_sink.h"
 #include <spdlog/details/log_msg.h>
 #include <spdlog/details/null_mutex.h>
-#include <spdlog/details/pattern_formatter.h>
+#include <spdlog/pattern_formatter.h>
 
 #include <algorithm>
 #include <memory>
