+++ src/spdlog.cpp
@@ -10,7 +10,7 @@
 #include <spdlog/details/backtracer-inl.h>
 #include <spdlog/details/registry-inl.h>
 #include <spdlog/details/os-inl.h>
-#include <spdlog/details/pattern_formatter-inl.h>
+#include <spdlog/pattern_formatter-inl.h>
 #include <spdlog/details/log_msg-inl.h>
 #include <spdlog/details/log_msg_buffer-inl.h>
 #include <spdlog/logger-inl.h>
